NIFI-6760: Fixed unit tests

+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/test/java/org/apache/nifi/controller/queue/clustered/server/TestStandardLoadBalanceProtocol.java
@@ -655,7 +655,7 @@ public class TestStandardLoadBalanceProtocol {
             final int length = Math.min(content.length - offset, 65535);
 
             out.write(DATA_FRAME_FOLLOWS);
-            out.writeShort(length);
+            out.writeInt(length);
             out.write(content, offset, length);
         }
